+++ src/Jobs/Job.php
@@ -3,7 +3,6 @@
 namespace Orchestra\Tenanti\Jobs;
 
 use Illuminate\Bus\Queueable;
-use Illuminate\Support\Facades\App;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Queue\InteractsWithQueue;
 
@@ -44,15 +43,29 @@ abstract class Job
      */
     protected function shouldBeFailed()
     {
-        if ($this->attempts() <= 3) {
-            return false;
-        }
-
-        if ($this->job) {
+        if ($this->attempts() > 3 && $this->job) {
             $this->job->failed();
+
+            return true;
         }
 
-        return true;
+        return false;
+    }
+
+    /**
+     * Should the job be delayed.
+     *
+     * @return bool
+     */
+    protected function shouldBeDelayed()
+    {
+        if ($this->job && is_null($this->model)) {
+            $this->job->release(10);
+
+            return true;
+        }
+
+        return false;
     }
 
     /**
@@ -62,6 +75,6 @@ abstract class Job
      */
     protected function resolveMigrator()
     {
-        return App::make('orchestra.tenanti')->driver($this->config['driver'] ?? null);
+        return resolve('orchestra.tenanti')->driver($this->config['driver'] ?? null);
     }
 }