+++ tests/Unit/Stubs/foo_migration.php
@@ -1,0 +1,6 @@
+<?php
+
+class FooMigration extends \Orchestra\Tenanti\Migration;
+{
+    //
+}