+++ tests/stubs/foo_migration.php
@@ -1,8 +1,0 @@
-<?php
-
-use Orchestra\Tenanti\Migration;
-
-class FooMigration extends Migration
-{
-    //
-}